feat: add membership Controller, Service, DTOs, Routes, Repository

+++ entity/tipe.go
@@ -1,22 +1,22 @@
 package entity
 
 type TipeMesin struct {
-	ID   uint `gorm:"primaryKey;autoIncrement"`
-	Tipe int  `gorm:"not null"`
+	ID   uint8 `gorm:"primaryKey;autoIncrement"`
+	Tipe int   `gorm:"not null"`
 
 	Timestamp
 }
 
 type TipeMobil struct {
-	ID   uint   `gorm:"primaryKey;autoIncrement"`
+	ID   uint8  `gorm:"primaryKey;autoIncrement"`
 	Tipe string `gorm:"not null"`
 
 	Timestamp
 }
 
 type TipePersneling struct {
-	ID   uint `gorm:"primaryKey;autoIncrement"`
-	Tipe int  `gorm:"not null"`
+	ID   uint8 `gorm:"primaryKey;autoIncrement"`
+	Tipe int   `gorm:"not null"`
 
 	Timestamp
 }
